backend/foodgram-st/requirements.txt
asgiref==3.6.0
certifi==2025.1.31
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
cryptography==44.0.2
defusedxml==0.7.1
Django==4.1.7
django-filter==21.1
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework_simplejwt==4.5.0
djoser==2.1.0
drf-extra-fields==3.7.0
filetype==1.2.0
flake8==7.2.0
idna==3.10
mccabe==0.7.0
oauthlib==3.2.2
pillow==11.1.0
psycopg2-binary==2.9.7
pycodestyle==2.13.0
pycparser==2.22
pyflakes==3.3.2
PyJWT==2.9.0
python-dotenv==1.0.1
python3-openid==3.2.0
reportlab==4.3.1
requests==2.32.3
requests-oauthlib==2.0.0
social-auth-app-django==4.0.0
social-auth-core==4.5.6
sqlparse==0.5.3
tzdata==2025.1
urllib3==1.26.15
